+++ src/pages/logbook/index.tsx
@@ -2,9 +2,7 @@ import { ReactElement } from 'react';
 import styled from 'styled-components';
 
 import { NextPageWithLayout } from '@/pages/_app';
-import AppLayout from '@/components/common/AppLayout/index';
-import Title from '@/components/common/Title';
-import TodoList from '@/components/common/TodoList/index';
+import { AppLayout, Title, TodoList } from '@/components/common';
 import { LOGBOOK } from '@/components/common/Figure';
 import { useLogbookTodoList } from '@/hooks/apis/todo/useTodoListQuery';
 import { useDeleteTodoMutation, useUpdateTodoMutation } from '@/hooks/apis/todo/useTodoMutation';
