+++ src/services/apis/todo.ts
@@ -3,7 +3,7 @@ import { v4 as uuidv4 } from 'uuid';
 import fetcher from '@/shared/utils/fetcher';
 import { syncTodos } from '@/shared/utils/sync';
 import { Todo, TodoCreateRequest, TodoUpdateRequest } from '@/shared/types/todo';
-import { todoStore } from '@/store/forage';
+import { todoStore } from '@/store/localForage/index';
 
 const todoService = {
   createTodo: async ({ locationName, longitude, latitude, dueDate }: TodoCreateRequest) =>
